+++ test/Tests/Wai/Route.hs
@@ -24,6 +24,8 @@ import Tests.Wai.Util
 
 import qualified Data.ByteString.Lazy as Lazy
 
+type ApplicationM m = Request -> (Response -> m ResponseReceived) -> m ResponseReceived
+
 tests :: TestTree
 tests = testGroup "Network.Wai.Routing"
     [ testCase "Sitemap" testSitemap
@@ -38,6 +40,7 @@ testSitemap = do
     ["/a", "/b", "/c", "/d", "/e", "/f", "/g", "/h"] @=? map fst routes
 
     let handler = route routes
+
     testEndpointA handler
     testEndpointB handler
     testEndpointC handler
@@ -48,7 +51,7 @@ testSitemap = do
 
 sitemap :: Routes Int IO ()
 sitemap = do
-    get "/a" handlerA $
+    get "/a" (continue handlerA) $
         accept "application" "json" .&. (query "name" .|. query "nick") .&. query "foo"
 
     attach 0
@@ -58,31 +61,31 @@ sitemap = do
 
     attach 1
 
-    get "/c" handlerC $
+    get "/c" (continue handlerC) $
         opt (query "foo")
 
     attach 2
 
-    get "/d" handlerD $
+    get "/d" (continue handlerD) $
         def 0 (query "foo")
 
     attach 3
 
-    get "/e" handlerE $
+    get "/e" (continue handlerE) $
         def 0 (header "foo")
 
     attach 4
 
-    get "/f" handlerF $
+    get "/f" (continue handlerF) $
         query "foo"
 
     attach 5
 
-    get "/g" handlerG true
+    get "/g" (continue handlerG) true
 
     attach 6
 
-    get "/h" handlerH $
+    get "/h" (continue handlerH) $
         cookie "user" .&. cookie "age"
 
     attach 7
@@ -90,8 +93,8 @@ sitemap = do
 handlerA :: Media "application" "json" ::: Int ::: ByteString -> IO Response
 handlerA (_ ::: i ::: _) = writeText (fromString . show $ i)
 
-handlerB :: Int -> IO Response
-handlerB baz = writeText (fromString . show $ baz)
+handlerB :: Int -> Continue IO -> IO ResponseReceived
+handlerB baz k = k $ responseLBS status200 [] (fromString . show $ baz)
 
 handlerC :: Maybe Int -> IO Response
 handlerC foo = writeText (fromString . show $ foo)
@@ -112,113 +115,113 @@ handlerH :: Lazy.ByteString ::: Int -> IO Response
 handlerH (user ::: age) = writeText $
     "user = " <> user <> ", age = " <> fromString (show age)
 
-testEndpointA :: Application -> Assertion
+testEndpointA :: ApplicationM IO -> Assertion
 testEndpointA f = do
     let rq = defaultRequest { rawPathInfo = "/a" }
 
-    rs0 <- f $ withHeader "Accept" "foo/bar" rq
+    rs0 <- apply f $ withHeader "Accept" "foo/bar" rq
     status406 @=? responseStatus rs0
 
-    rs1 <- f $ json rq
+    rs1 <- apply f $ json rq
     status400 @=? responseStatus rs1
 
-    rs2 <- f . json . withQuery "name" "x" $ rq
+    rs2 <- apply f $ json . withQuery "name" "x" $ rq
     status400 @=? responseStatus rs2
 
-    rs3 <- f . json . withQuery "name" "123" . withQuery "foo" "\"z\"" $ rq
+    rs3 <- apply f $ json . withQuery "name" "123" . withQuery "foo" "\"z\"" $ rq
     status200 @=? responseStatus rs3
 
 
-testEndpointB :: Application -> Assertion
+testEndpointB :: ApplicationM IO -> Assertion
 testEndpointB f = do
     let rq = defaultRequest { rawPathInfo = "/b" }
 
-    rs0 <- f rq
+    rs0 <- apply f rq
     status400 @=? responseStatus rs0
     "'baz' not-available [query]" @=? responseBody rs0
 
-    rs1 <- f . withQuery "baz" "abc" $ rq
+    rs1 <- apply f $ withQuery "baz" "abc" $ rq
     status400 @=? responseStatus rs1
     "'baz' type-error [query] -- Failed reading: Invalid Int" @=? responseBody rs1
 
-    rs2 <- f . withQuery "baz" "abc" . withQuery "baz" "123" $ rq
+    rs2 <- apply f $ withQuery "baz" "abc" . withQuery "baz" "123" $ rq
     status200 @=? responseStatus rs2
     "123" @=? responseBody rs2
 
 
-testEndpointC :: Application -> Assertion
+testEndpointC :: ApplicationM IO -> Assertion
 testEndpointC f = do
     let rq = defaultRequest { rawPathInfo = "/c" }
 
-    rs0 <- f rq
+    rs0 <- apply f rq
     status200 @=? responseStatus rs0
     "Nothing" @=? responseBody rs0
 
-    rs1 <- f . withQuery "foo" "abc" . withQuery "foo" "123" $ rq
+    rs1 <- apply f $ withQuery "foo" "abc" . withQuery "foo" "123" $ rq
     status200  @=? responseStatus rs1
     "Just 123" @=? responseBody rs1
 
-    rs2 <- f . withQuery "foo" "abc" $ rq
+    rs2 <- apply f $ withQuery "foo" "abc" $ rq
     status400 @=? responseStatus rs2
     "'foo' type-error [query] -- Failed reading: Invalid Int" @=? responseBody rs2
 
 
-testEndpointD :: Application -> Assertion
+testEndpointD :: ApplicationM IO -> Assertion
 testEndpointD f = do
     let rq = defaultRequest { rawPathInfo = "/d" }
 
-    rs0 <- f rq
+    rs0 <- apply f rq
     status200 @=? responseStatus rs0
     "0"       @=? responseBody rs0
 
-    rs1 <- f . withQuery "foo" "xxx" . withQuery "foo" "42" $ rq
+    rs1 <- apply f $ withQuery "foo" "xxx" . withQuery "foo" "42" $ rq
     status200 @=? responseStatus rs1
     "42"      @=? responseBody rs1
 
-    rs2 <- f . withQuery "foo" "yyy" $ rq
+    rs2 <- apply f $ withQuery "foo" "yyy" $ rq
     status400 @=? responseStatus rs2
     "'foo' type-error [query] -- Failed reading: Invalid Int" @=? responseBody rs2
 
 
-testEndpointE :: Application -> Assertion
+testEndpointE :: ApplicationM IO -> Assertion
 testEndpointE f = do
     let rq = defaultRequest { rawPathInfo = "/e" }
 
-    rs0 <- f rq
+    rs0 <- apply f rq
     status200 @=? responseStatus rs0
     "0"       @=? responseBody rs0
 
-    rs1 <- f $ withHeader "foo" "42" rq
+    rs1 <- apply f $ withHeader "foo" "42" rq
     status200 @=? responseStatus rs1
     "42"      @=? responseBody rs1
 
-    rs2 <- f $ withHeader "foo" "abc" rq
+    rs2 <- apply f $ withHeader "foo" "abc" rq
     status400 @=? responseStatus rs2
     "'foo' type-error [header] -- Failed reading: Invalid Int" @=? responseBody rs2
 
 
-testEndpointF :: Application -> Assertion
+testEndpointF :: ApplicationM IO -> Assertion
 testEndpointF f = do
     let rq = defaultRequest { rawPathInfo = "/f" }
 
-    rs0 <- f . withQuery "foo" "1,2,3,4" $ rq
+    rs0 <- apply f $ withQuery "foo" "1,2,3,4" $ rq
     status200 @=? responseStatus rs0
     "10"      @=? responseBody rs0
 
 
-testEndpointH :: Application -> Assertion
+testEndpointH :: ApplicationM IO -> Assertion
 testEndpointH f = do
     let rq = defaultRequest { rawPathInfo = "/h" }
 
-    rs0 <- f rq
+    rs0 <- apply f rq
     status400 @=? responseStatus rs0
     "'user' not-available [cookie]" @=? responseBody rs0
 
-    rs1 <- f . withHeader "Cookie" "user=joe" $ rq
+    rs1 <- apply f $ withHeader "Cookie" "user=joe" $ rq
     status400 @=? responseStatus rs1
     "'age' not-available [cookie]" @=? responseBody rs1
 
-    rs2 <- f . withHeader "Cookie" "user=joe; age=42" $ rq
+    rs2 <- apply f $ withHeader "Cookie" "user=joe; age=42" $ rq
     status200 @=? responseStatus rs2
     "user = joe, age = 42" @=? responseBody rs2
 
@@ -233,8 +236,8 @@ testMedia = do
 
 sitemapMedia :: Routes a IO ()
 sitemapMedia = do
-    get "/media" handlerJson   $ accept "application" "json"
-    get "/media" handlerThrift $ accept "application" "x-thrift"
+    get "/media" (continue handlerJson)   $ accept "application" "json"
+    get "/media" (continue handlerThrift) $ accept "application" "x-thrift"
 
 handlerJson :: Media "application" "json" -> IO Response
 handlerJson _ = writeText "application/json"
@@ -242,8 +245,8 @@ handlerJson _ = writeText "application/json"
 handlerThrift :: Media "application" "x-thrift" -> IO Response
 handlerThrift _ = writeText "application/x-thrift"
 
-expectMedia :: ByteString -> ByteString -> (RoutingReq -> IO Response) -> Assertion
+expectMedia :: ByteString -> ByteString -> App IO -> Assertion
 expectMedia h res m = do
     let rq = defaultRequest { rawPathInfo = "/media" }
-    rs <- m . fromReq [] . fromRequest . withHeader "Accept" h $ rq
+    rs <- apply m $ fromReq [] . fromRequest . withHeader "Accept" h $ rq
     Lazy.fromStrict res @=? responseBody rs